fixes isPortReady as a separate module, adds function names and lowercases TTL
